0.1: Changes
Going to start marking these commits with numbers starting today.

- cards.java simply had its main method commented out since it was only
used for debugging that class.
- deck.java had its debugging code updated to reflect the abstractino of
the for-loop which printed all the cards in a deck. last_modified was
updated to reflect this change.
- emptyDeck.java has a new method called moveCardWithLocation which
moves a card from a specific location from a deck to another deck. This
is different from just moving a card which just grabs a card from the
top of a deck and adds it to the top of another deck.
showHand() is an abstraction of some previous debugging code which
prints to the console all the cards in a deck.
- gameFunc.java has a new function placeDownCards() which handles
players placing down cards on top of their facedown cards. New variable
handles the limit for that method.
- game.java has some new code for generating who puts their card down
first.